requirements.txt
matplotlib==3.9.0
pandas==2.2.2
pillow==10.3.0
requests==2.32.2
seaborn==0.13.2
selenium==4.21.0
streamlit==1.35.0
